rename core module to shared
