+++ app/src/main/java/com/me/recipe/ui/MainActivity.kt
@@ -3,8 +3,8 @@ package com.me.recipe.ui
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.fragment.app.FragmentActivity
-import com.me.recipe.core.datastore.SettingsDataStore
-import com.me.recipe.core.datastore.UserDataStore
+import com.me.recipe.shared.datastore.SettingsDataStore
+import com.me.recipe.shared.datastore.UserDataStore
 import com.me.recipe.ui.theme.RecipeTheme
 import com.me.recipe.util.extention.getPushNotificationPermission
 import dagger.hilt.android.AndroidEntryPoint
