+++ app/src/main/java/com/me/recipe/ui/recipe/RecipeViewModel.kt
@@ -3,7 +3,7 @@ package com.me.recipe.ui.recipe
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.me.recipe.core.utils.TAG
+import com.me.recipe.shared.utils.TAG
 import com.me.recipe.ui.navigation.RecipeDestination
 import dagger.Lazy
 import dagger.hilt.android.lifecycle.HiltViewModel
