+++ cache/build.gradle.kts
@@ -46,7 +46,7 @@ spotless {
 }
 
 dependencies {
-    implementation(project(":core"))
+    implementation(project(":shared"))
     implementation(project(":domain"))
 
     implementation(libs.hilt)
