+++ data/src/main/java/com/me/recipe/data/features/recipe/mapper/RecipeEntityMapper.kt
@@ -1,7 +1,7 @@
 package com.me.recipe.data.features.recipe.mapper
 
 import com.me.recipe.cache.recipe.model.RecipeEntity
-import com.me.recipe.core.utils.DateUtils
+import com.me.recipe.shared.utils.DateUtils
 import com.me.recipe.domain.features.recipe.model.Recipe
 import com.me.recipe.domain.util.DomainMapper
 import kotlinx.collections.immutable.toPersistentList
