+++ data/src/main/java/com/me/recipe/data/features/recipelist/repository/RecipeListRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.me.recipe.data.features.recipelist.repository
 
 import com.me.recipe.cache.recipe.RecipeDao
-import com.me.recipe.core.data.DataState
-import com.me.recipe.core.utils.RECIPE_PAGINATION_PAGE_SIZE
+import com.me.recipe.shared.data.DataState
+import com.me.recipe.shared.utils.RECIPE_PAGINATION_PAGE_SIZE
 import com.me.recipe.data.features.recipe.mapper.RecipeDtoMapper
 import com.me.recipe.data.features.recipe.mapper.RecipeEntityMapper
 import com.me.recipe.domain.features.recipe.model.Recipe
