+++ domain/src/main/java/com/me/recipe/domain/features/recipe/repository/RecipeRepository.kt
@@ -1,6 +1,6 @@
 package com.me.recipe.domain.features.recipe.repository
 
-import com.me.recipe.core.data.DataState
+import com.me.recipe.shared.data.DataState
 import com.me.recipe.domain.features.recipe.model.Recipe
 import kotlinx.coroutines.flow.Flow
