+++ domain/src/main/java/com/me/recipe/domain/features/recipe/usecases/GetRecipeUsecase.kt
@@ -1,6 +1,6 @@
 package com.me.recipe.domain.features.recipe.usecases
 
-import com.me.recipe.core.data.DataState
+import com.me.recipe.shared.data.DataState
 import com.me.recipe.domain.features.recipe.model.Recipe
 import com.me.recipe.domain.features.recipe.repository.RecipeRepository
 import javax.inject.Inject
