+++ domain/src/main/java/com/me/recipe/domain/features/recipelist/usecases/SearchRecipesUsecase.kt
@@ -1,6 +1,6 @@
 package com.me.recipe.domain.features.recipelist.usecases
 
-import com.me.recipe.core.data.DataState
+import com.me.recipe.shared.data.DataState
 import com.me.recipe.domain.features.recipe.model.Recipe
 import com.me.recipe.domain.features.recipelist.repository.RecipeListRepository
 import javax.inject.Inject
