+++ network/src/main/java/com/me/recipe/network/core/di/retrofit/InterceptorModule.kt
@@ -1,6 +1,6 @@
 package com.me.recipe.network.core.di.retrofit
 
-import com.me.recipe.core.datastore.UserDataStore
+import com.me.recipe.shared.datastore.UserDataStore
 import com.me.recipe.network.core.interceptors.AuthenticationInterceptor
 import dagger.Module
 import dagger.Provides
