+++ settings.gradle
@@ -18,7 +18,7 @@ dependencyResolutionManagement {
 rootProject.name = "Recipe"
 include ':app'
 include ':network'
-include ':core'
+include ':shared'
 include ':domain'
 include ':cache'
 include ':data'
