+++ shared/src/main/AndroidManifest.xml
