+++ shared/src/main/java/com/me/recipe/shared/datastore/SettingsDataStore.kt
@@ -1,12 +1,13 @@
-package com.me.recipe.core.datastore
+package com.me.recipe.shared.datastore
 
 import android.content.Context
 import androidx.compose.runtime.mutableStateOf
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
+import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlinx.coroutines.CoroutineScope
@@ -16,7 +17,9 @@ import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 
 @Singleton
-class UserDataStore @Inject constructor(val context: Context) {
+class SettingsDataStore
+@Inject
+constructor(@ApplicationContext val context: Context) {
 
     private val scope = CoroutineScope(Main)
 
@@ -24,28 +27,29 @@ class UserDataStore @Inject constructor(val context: Context) {
         observeDataStore()
     }
 
-    val accessToken = mutableStateOf("")
+    val isDark = mutableStateOf(false)
 
-    fun setAccessToken(accessToken: String) {
+    fun toggleTheme() {
         scope.launch {
             context.dataStore.edit { preferences ->
-                preferences[ACCESS_TOKEN_KEY] = accessToken
+                val current = preferences[DARK_THEME_KEY] ?: false
+                preferences[DARK_THEME_KEY] = !current
             }
         }
     }
 
     private fun observeDataStore() {
         context.dataStore.data.onEach { preferences ->
-            preferences[ACCESS_TOKEN_KEY]?.let { token ->
-                accessToken.value = token
+            preferences[DARK_THEME_KEY]?.let { isDarkTheme ->
+                isDark.value = isDarkTheme
             }
         }.launchIn(scope)
     }
 
     companion object {
-        private val ACCESS_TOKEN_KEY = stringPreferencesKey("access_token_key")
+        private val DARK_THEME_KEY = booleanPreferencesKey("dark_theme_key")
 
-        private const val APP_PREFERENCE_NAME = "user_info"
+        private const val APP_PREFERENCE_NAME = "settings"
         private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(
             name = APP_PREFERENCE_NAME,
         )
